+++ App/migrations/20230102100648-create-emission-region-static.js
@@ -23,7 +23,7 @@ module.exports = {
         type: Sequelize.DATE
       },
       intensity: {
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL(10,1)
       },
       createdAt: {
         allowNull: false,
