+++ App/migrations/20230102130430-create-region-emission-static.js
@@ -26,7 +26,7 @@ module.exports = {
         type: Sequelize.DATE
       },
       contributor: {
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL(10,1)
       },
       createdAt: {
         allowNull: false,
